+++ src/Internal/Helper.php
@@ -45,4 +45,48 @@ final class Helper
     {
         return \is_string($field) && $field !== '' || \is_int($field);
     }
+    
+    /**
+     * @param mixed $value
+     * @return string
+     */
+    public static function describe($value): string
+    {
+        if ($value === null) {
+            return 'NULL';
+        }
+        
+        if (\is_bool($value)) {
+            return $value ? 'TRUE' : 'FALSE';
+        }
+        
+        if (\is_numeric($value)) {
+            return (string) $value;
+        }
+        
+        if (\is_string($value)) {
+            return \mb_strlen($value) > 50 ? \mb_substr($value, 0, 47).'...' : $value;
+        }
+        
+        if (\is_array($value)) {
+            $desc = 'array of length: '.\count($value);
+            
+            $json = \json_encode($value);
+            if ($json !== false) {
+                if (\mb_strlen($json) > 50) {
+                    $desc .= ' '.\mb_substr($json, 0, 47).'...';
+                } else {
+                    $desc .= ' '.$json;
+                }
+            }
+            
+            return $desc;
+        }
+        
+        if (\is_object($value)) {
+            return 'object of class: '.\get_class($value);
+        }
+        
+        return \gettype($value);
+    }
 }